Add files via upload
This zip folder includes the final models for each country / crop combination. It also includes a file called "modelsRanked.csv", which shows each type of model and the count of times it was considered the "best" model for a country /crop combination based on its R^2 value. Finally, there is a file called "allModelsDF.csv", which has columns: country, crop, modelType, and R^2.